Move local type tests to their related tests
