+++ src/test/java/spoon/test/SpoonTestHelpers.java
@@ -16,14 +16,19 @@
  */
 package spoon.test;
 
+import spoon.Launcher;
 import spoon.metamodel.Metamodel;
+import spoon.reflect.CtModel;
+import spoon.reflect.code.CtBlock;
 import spoon.reflect.declaration.CtElement;
 import spoon.reflect.declaration.CtMethod;
 import spoon.reflect.declaration.CtType;
 import spoon.reflect.reference.CtTypeReference;
 import spoon.reflect.visitor.filter.OverridingMethodFilter;
+import spoon.reflect.visitor.filter.TypeFilter;
 import spoon.support.DerivedProperty;
 import spoon.support.UnsettableProperty;
+import spoon.support.compiler.VirtualFile;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -150,5 +155,32 @@ public class SpoonTestHelpers {
 	/** Place at the top of a JUnit4 test method to disable it for Windows. */
 	public static void assumeNotWindows() {
 		assumeFalse(System.getProperty("os.name").toLowerCase().contains("windows"));
+	}
+
+	/** Builds a model from code given as string */
+	public static CtModel createModelFromString(String code, int complianceLevel) {
+		Launcher launcher = new Launcher();
+		launcher.getEnvironment().setComplianceLevel(complianceLevel);
+		launcher.getEnvironment().setNoClasspath(true);
+		launcher.addInputResource(new VirtualFile(code));
+		return launcher.buildModel();
+	}
+
+	/** Returns the block of the first non-implicit method */
+	public static CtBlock<?> getBlock(CtModel model) {
+		return model.getElements(new TypeFilter<>(CtBlock.class))
+				.stream()
+				.filter(b -> b.getParent() instanceof CtMethod<?> && !b.getParent().isImplicit())
+				.findFirst()
+				.orElseThrow(() -> new IllegalArgumentException("code does not contain explicit method"));
+	}
+
+	/** wraps code into a method in a class */
+	public static String wrapLocal(String localDeclarationSnippet) {
+		return  "class X {\n" +
+				"	public void myMethod() {\n" +
+				localDeclarationSnippet +
+				"	}\n" +
+				"}";
 	}
 }
